Refactor: move `_BaseTestWithGui` to `testing` module
