+++ qt_binder/tests/test_raw_widgets.py
@@ -15,23 +15,10 @@
 
 import unittest
 
-from pyface.ui.qt4.util.gui_test_assistant import GuiTestAssistant
-from traits.api import pop_exception_handler, push_exception_handler
-
 from ..qt import QtCore, QtGui
 from ..raw_widgets import BasicGridLayout, GroupBox, HBoxLayout, Label, \
     LineEdit, Object, VBoxLayout, Widget, binder_registry
-
-
-class _BaseTestWithGui(GuiTestAssistant):
-
-    def setUp(self):
-        GuiTestAssistant.setUp(self)
-        push_exception_handler(reraise_exceptions=True)
-
-    def tearDown(self):
-        pop_exception_handler()
-        GuiTestAssistant.tearDown(self)
+from ..testing import _BaseTestWithGui
 
 
 class TestBoxLayout(_BaseTestWithGui, unittest.TestCase):
